Import of INA/Char-UHC-0.93 from CPAN.
gitpan-cpan-distribution: Char-UHC
gitpan-cpan-version:      0.93
gitpan-cpan-path:         INA/Char/UHC/Char-UHC-0.93.tar.gz
gitpan-cpan-author:       INA
gitpan-cpan-maturity:     released

+++ t/234_rmdir.t
@@ -12,6 +12,12 @@ if ($^O !~ /\A (?: MSWin32 | NetWare | symbian | dos ) \z/oxms) {
     exit;
 }
 
+if ($ENV{'PERL5SHELL'} =~ /Win95Cmd\.exe/xmsi) {
+    print "ok - 1 # SKIP $^X $0\n";
+    print "ok - 2 # SKIP $^X $0\n";
+    exit;
+}
+
 mkdir('directory',0777);
 if (rmdir('directory')) {
     print "ok - 1 rmdir $^X $__FILE__\n";
